refac(): Rename function getEmailData to getUserData
